core/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
Django==4.2.17
django-debug-toolbar==5.0.1
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
kombu==5.4.2
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.0
six==1.17.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
